fix: split build and compile into multiple files
