+++ src/compile/compile-string.ts
@@ -1,0 +1,7 @@
+import { wrapImplementation } from '../compiler';
+
+export function sassCompileString( source: string, options?: CliOptions ) : string {
+    const compiler = wrapImplementation( options );
+
+    return compiler.compileString( source, options?.sassOptions );
+}
