+++ src/compile/compile.ts
@@ -1,0 +1,7 @@
+import { wrapImplementation } from '../compiler';
+
+export function sassCompile( file: string, options?: CliOptions ) : string {
+    const compiler = wrapImplementation( options );
+
+    return compiler.compile( file, options?.sassOptions );
+}
