+++ src/index.ts
@@ -1,7 +1,7 @@
 /// <reference path="../types/index.d.ts" />
 
-export { sassBuild, sassBuildFiles, sassBuildString } from './sass-build';
-export { sassCompile, sassCompileString } from './sass-compile';
+export { sassBuild, sassBuildFiles, sassBuildString } from './build';
+export { sassCompile, sassCompileString } from './compile';
 export { wrapImplementation as wrapCompiler } from './compiler';
 export { sassCacheImporter } from './importers/cache-importer';
 export { sassPackageImporter } from './importers/package-importer';
