--- src/sass-compile.ts
@@ -1,18 +1,0 @@
-import { wrapImplementation } from './compiler';
-
-function sassCompile( file: string, options?: CliOptions ) : string {
-    const compiler = wrapImplementation( options );
-
-    return compiler.compile( file, options?.sassOptions );
-}
-
-function sassCompileString( source: string, options?: CliOptions ) : string {
-    const compiler = wrapImplementation( options );
-
-    return compiler.compileString( source, options?.sassOptions );
-}
-
-export {
-    sassCompile,
-    sassCompileString
-};
